+++ main.cpp
@@ -1,0 +1,37 @@
+#include "sudokuGen.hpp"
+
+using namespace std;
+
+// START: The main function
+int main(int argc, char const *argv[])
+{
+  // Initialising seed for random number generation
+  srand(time(NULL));
+
+  // Creating an instance of Sudoku
+  Sudoku *puzzle = new Sudoku();
+
+  // Creating a seed for puzzle generation
+  puzzle->createSeed();
+
+  // Generating the puzzle
+  puzzle->genPuzzle();
+
+  // Calculating difficulty of puzzle
+  puzzle->calculateDifficulty();
+
+  // testing by printing the grid
+  puzzle->printGrid();
+
+  // Printing the grid into SVG file
+  string rem = "sudokuGen";
+  string path = argv[0];
+  path = path.substr(0,path.size() - rem.size());
+  puzzle->printSVG(path);
+  cout<<"The above sudoku puzzle has been stored in puzzles.svg in current folder\n";
+  // freeing the memory
+  delete puzzle;
+
+  return 0;
+}
+// END: The main function
